requirements.txt
# Core
langgraph>=0.0.26
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
pydantic>=2.6.0

# Database & Vector Store
qdrant-client>=1.15.0
# FHIR / MCP
fhir.resources>=7.1.0
mcp>=0.1.0

# API / Web
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
httpx>=0.27.0

# Observability
arize-phoenix>=3.0.0
ragas>=0.1.0

# Data Processing
pandas>=2.2.0
tiktoken>=0.6.0
PyPDF2>=3.0.0

